pyproject.toml
[project]
name = "etf-agent"
version = "0.1.0"
description = "ETF 및 주식 종목 데이터 분석 에이전트"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "semantic-kernel>=1.14.0",
    "finnhub-python>=2.4.0",
    "azure-cosmos>=4.7.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "azure-core-tracing-opentelemetry<=1.0.0b12",
    "azure-monitor-opentelemetry>=1.6.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b26",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-openai>=0.29.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]
